refactor: Refactor cms modules.
