+++ app/.server/cms/index.ts
@@ -1,4 +1,5 @@
-export * from './api';
+import type { TagResponse } from './type';
+
+export type { TagResponse };
 export * from './client';
-export * from './schema';
-export * from './type';
+export * from './api';
